JSDK-252 【点播】查询视频信息,代码审查20210310

+++ polyv-java-vod-sdk/src/main/java/net/polyv/vod/v1/entity/subaccount/query/VodQueryVideoInfoResponse.java
@@ -20,7 +20,7 @@ import net.polyv.vod.v1.entity.VodCommonResponse;
 @EqualsAndHashCode(callSuper = true)
 @Accessors(chain = true)
 @ApiModel("查询视频信息返回实体")
-public class VodQueryVideoInfoResponse extends VodCommonResponse{
+public class VodQueryVideoInfoResponse extends VodCommonResponse {
     
     /**
      * 视频id
@@ -56,7 +56,7 @@ public class VodQueryVideoInfoResponse extends VodCommonResponse{
     @Data
     @Accessors(chain = true)
     @ApiModel("查询的视频截图")
-    public static class SnapshotInfo{
+    public static class SnapshotInfo {
         /**
          * 截图url数组
          */
@@ -67,49 +67,49 @@ public class VodQueryVideoInfoResponse extends VodCommonResponse{
     @Data
     @Accessors(chain = true)
     @ApiModel("查询的视频元数据")
-    public static class MetaData{
+    public static class MetaData {
         /**
          * 源文件大小，单位：Bytes
          */
         @ApiModelProperty(name = "size", value = "源文件大小，单位：Bytes", required = false)
-        private Integer size;
-    
+        private Long size;
+        
         /**
          * 视频容器类型，如mp4、flv等
          */
         @ApiModelProperty(name = "format", value = "视频容器类型，如mp4、flv等", required = false)
         private String format;
-    
+        
         /**
          * 源视频时长，单位：秒
          */
         @ApiModelProperty(name = "duration", value = "源视频时长，单位：秒", required = false)
-        private Float duration;
-    
+        private Integer duration;
+        
         /**
          * 视频码率，单位：bps
          */
         @ApiModelProperty(name = "bitrate", value = "视频码率，单位：bps", required = false)
         private Integer bitrate;
-    
+        
         /**
          * 视频帧率
          */
         @ApiModelProperty(name = "fps", value = "视频帧率", required = false)
         private Integer fps;
-    
+        
         /**
          * 分辨率高，单位：px
          */
         @ApiModelProperty(name = "height", value = "分辨率高，单位：px", required = false)
         private Integer height;
-    
+        
         /**
          * 分辨率宽，单位：px
          */
         @ApiModelProperty(name = "width", value = "分辨率宽，单位：px", required = false)
         private Integer width;
-    
+        
         /**
          * 编码格式，如h264、h265等
          */
@@ -120,61 +120,61 @@ public class VodQueryVideoInfoResponse extends VodCommonResponse{
     @Data
     @Accessors(chain = true)
     @ApiModel("查询的视频转码信息")
-    public static class TranscodeInfos{
+    public static class TranscodeInfos {
         /**
          * 播放地址
          */
         @ApiModelProperty(name = "playUrl", value = "播放地址", required = false)
         private String playUrl;
-    
+        
         /**
          * 清晰度，SOURCE:原清晰度,LD:普清,SD:标清,HD:高清
          */
         @ApiModelProperty(name = "definition", value = "清晰度，SOURCE:原清晰度,LD:普清,SD:标清,HD:高清", required = false)
         private String definition;
-    
+        
         /**
          * 时长，秒
          */
         @ApiModelProperty(name = "duration", value = "时长，秒", required = false)
         private Integer duration;
-    
+        
         /**
          * 加密视频为1，非加密为0
          */
         @ApiModelProperty(name = "encrypt", value = "加密视频为1，非加密为0", required = false)
         private Boolean encrypt;
-    
+        
         /**
          * 转码格式，如mp4、flv、pdx、hls
          */
         @ApiModelProperty(name = "format", value = "转码格式，如mp4、flv、pdx、hls", required = false)
         private String format;
-    
+        
         /**
          * 视频帧率
          */
         @ApiModelProperty(name = "fps", value = "视频帧率", required = false)
         private Integer fps;
-    
+        
         /**
          * 码率kbps
          */
         @ApiModelProperty(name = "bitrate", value = "码率kbps", required = false)
         private Integer bitrate;
-    
+        
         /**
          * 分辨率高，单位：px
          */
         @ApiModelProperty(name = "height", value = "分辨率高，单位：px", required = false)
         private Integer height;
-    
+        
         /**
          * 分辨率宽，单位：px
          */
         @ApiModelProperty(name = "width", value = "分辨率宽，单位：px", required = false)
         private Integer width;
-    
+        
         /**
          * 视频状态, normal:可以正常播放,unavailable:不能正常播放
          */
@@ -203,7 +203,7 @@ public class VodQueryVideoInfoResponse extends VodCommonResponse{
          * 源视频时长，单位：秒
          */
         @ApiModelProperty(name = "duration", value = "源视频时长，单位：秒", required = false)
-        private Float duration;
+        private Integer duration;
         
         /**
          * 首图地址，大图
@@ -234,7 +234,7 @@ public class VodQueryVideoInfoResponse extends VodCommonResponse{
          */
         @ApiModelProperty(name = "status", value = "视频状态码;60/61:已发布;10:等待编码;20:正在编码;50:等待审核;51:审核不通过;-1:已删除;",
                 required = false)
-        private Long status;
+        private Integer status;
         
         /**
          * 分类id, 如1为根目录
